honor fastboot 1.0.0 format over browser intitializer
